fix example type cast

+++ examples/state-checking/state-checking.c
@@ -116,7 +116,7 @@ int main(void) {
             printf("Is middle mouse button released: %s\n", currState.middleMouseReleased ? "Yes" : "No");
         }
         if (currState.scrollX != prevState.scrollX || currState.scrollY != prevState.scrollY) {
-            printf("Mouse scrolling (%f %f)\n", currState.scrollX, currState.scrollY);
+            printf("Mouse scrolling (%f %f)\n", (float)currState.scrollX, (float)currState.scrollY);
         }
         if (RGFW_isKeyDown(RGFW_controlL) && (currState.mouseX != prevState.mouseX || currState.mouseY != prevState.mouseY)) {
             printf("Mouse position in window: (%i, %i)\n", currState.mouseX, currState.mouseY);
